--- includes/default.hpp
@@ -1,20 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   default.hpp                                        :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mmoussou <mmoussou@student.42angouleme.fr  +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/02/11 13:42:04 by mmoussou          #+#    #+#             */
-/*   Updated: 2025/02/11 13:43:13 by mmoussou         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#pragma once
-
-#ifndef __WEBSERV_DEFAULT_HPP__
-# define __WEBSERV_DEFAULT_HPP__
-
-using namespace webserv; // oh no, cringo D:
-
-#endif //__WEBSERV_DEFAULT_HPP__
