+++ includes/server/default.hpp
@@ -6,15 +6,13 @@
 /*   By: mmoussou <mmoussou@student.42angouleme.fr  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/11 13:29:05 by mmoussou          #+#    #+#             */
-/*   Updated: 2025/02/11 13:43:26 by mmoussou         ###   ########.fr       */
+/*   Updated: 2025/02/11 15:00:15 by mmoussou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #pragma once
 #ifndef __WEBSERV_SERVER_DEFAULT_HPP__
 # define __WEBSERV_SERVER_DEFAULT_HPP__
-
-#include "default.hpp"
 
 namespace webserv {
 namespace server {
